Create ERC1155ABOSCH.sol
A simple template for artists who want to mint multiple NFT's with a single contract, batch transfer capability.  With this contract your tokens will be seen by ERC-721 platforms and wallets that haven't yet implemented ERC-1155 protocol.  Just swap out the variables at the bottom of the contract, compile and deploy :)  As always, test on test-network before deploying to Mainnet.